+++ state_beacon/test/src/beacons/readable_test.dart
@@ -1,0 +1,44 @@
+import 'dart:async';
+
+import 'package:flutter_test/flutter_test.dart';
+import 'package:state_beacon/state_beacon.dart';
+
+void main() {
+  test('should set initial value', () {
+    var beacon = Beacon.readable(10);
+    expect(beacon.peek(), equals(10));
+  });
+
+  test('should decrease listenersCount when unsubscribed', () {
+    var beacon = Beacon.readable(10);
+
+    final unsub1 = beacon.subscribe((_) {});
+
+    expect(beacon.listenersCount, 1);
+
+    final unsub2 = Beacon.createEffect(() {
+      beacon.value;
+    });
+
+    expect(beacon.listenersCount, 2);
+
+    unsub1();
+
+    expect(beacon.listenersCount, 1);
+
+    unsub2();
+
+    expect(beacon.listenersCount, 0);
+  });
+
+  test('should fire subscription immediately', () async {
+    final a = Beacon.readable(1);
+    final completer = Completer<int>();
+
+    a.subscribe(completer.complete, startNow: true);
+
+    final result = await completer.future;
+
+    expect(result, 1);
+  });
+}
